% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.R
\docType{data}
\name{demo}
\alias{demo}
\title{Demo data.}
\format{
A data frame with 4 rows and 2 variables.
\describe{
\item{first_name}{The given name of the data hunter.}
\item{last_name}{The family name of the data hunter.}
}
}
\source{
N.Paterno
}
\usage{
demo
}
\description{
A demo dataset consisting of names of some of the data hunters.
}
\examples{
 unique(demo$last_name)

}
\keyword{datasets}
