% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-peanut_allergy.R
\docType{data}
\name{peanut_allergy}
\alias{peanut_allergy}
\title{Peanut Allergy Experiment in Young Children}
\format{
A data frame with 30 rows and 2 variables.
\describe{
\item{had_early_risk}{Indicates if the child was deemed to
be at elevated risk of developing a peanut allergy at the start
of the study based on a pin-prick test.}
\item{regimen}{The peanut regimen assigned for the child,
either "avoid" or "consume".}
\item{allergic}{Indicates if the child had developed a peanut
allergy by the end of the study.}
}
}
\source{
\href{https://www.nejm.org/doi/full/10.1056/NEJMoa1414850}
{Du Toit G, Roberts G, et al. 2015. Randomized Trial of
Peanut Consumption in Infants at Risk for Peanut Allergy.
New England Journal of Medicine 372:803-813.}

See also: \href{https://www.youtube.com/watch?v=IRUisFcZ2gk}
{The Facts about Peanut Allergies Might Surprise You
(Healthcare Triage episode)}.
}
\usage{
peanut_allergy
}
\description{
In the late 1990s, it was believed that young children should
exclude peanuts from their diets to reduce the chance of developing
an allergy. In 2008, researchers were no longer so sure. This
experiment evaluates whether peanut exposure is helpful or harmful.
A peanut diet regimen (consume or avoid) was assigned to over 500
young children during years 2-5, ages during which children had
previously been told not to eat peanuts. The key outcome was testing
for a peanut allergy when each child turned 5.
}
\examples{
table(peanut_allergy[c("regimen", "allergic", "had_early_risk")])

}
\keyword{datasets}
